+++ fiddles/chef/kitchen/fiddle-0000-Template/recipes/default.rb
@@ -1,0 +1,10 @@
+#
+# Cookbook:: fiddle-0000-Template
+# Recipe:: default
+#
+# Copyright:: 2018, The Authors, All Rights Reserved.
+
+cookbook_file "/etc/motd" do
+  source "motd"
+  mode "0644"
+end