+++ fiddles/chef/kitchen/fiddle-0001-Apache/recipes/default.rb
@@ -1,0 +1,5 @@
+#
+# Cookbook:: fiddle-0001-Apache
+# Recipe:: default
+#
+# Copyright:: 2018, The Authors, All Rights Reserved.